refactor(shared): split CLI runner helpers
