+++ packages/shared/src/cli/index.ts
@@ -1,9 +1,4 @@
-export {
-  runToolsCLI,
-  CLIError,
-  reportCLIError,
-  parseValue,
-  parseCliArgs,
-  removePrefix,
-} from './cli-runner';
+export { CLIError, reportCLIError } from './cli-error';
+export { parseCliArgs, parseValue } from './cli-args';
+export { runToolsCLI, removePrefix } from './cli-runner';
 export type { CLIRunnerOptions, CLIExtraCommand } from './cli-runner';
